fix: Update link button color in Alert component to be consistent
In Alert, link color should be always consistent regardless of the
link variants. Before this change, the button-link inside an Alert
had the color of a button, not the one of a link.

This got fixed by introducing a new design token and not using the
`colorTextButtonNormalActive` token for the link button color anymore.

To have this scenario tested in screenshots, I've added additional links to the Alert component testing page.