fix(date-picker): Fix baseDate handling when navigating through the calendar
This change fixes two issues:

1) Scenario value set to a day in March, all dates <= April are disabled:
When opening the date picker, April was shown as month
because and not march as the value stated. With this change, March is
now shown instead of April to reflect the current input value.

2) When attempting to use isDateEnabled given a set of available dates,
there was an issue when the list of available dates was not continuous
by month. E.g. if the following dates `['2022-04-14', '2022-03-14', '2022-01-14']`
are enabled, the component was not allowing to go back to previous months:
navigating from March 2022 into the past was not possible (making 2022-01-14 inaccessible).