chore: Integrate d3-scale as bundled dependency, upgrade d3-scale (#399)
With this change we add d3-scale as bundled dependency. This is because the recent d3-scale version bump (#371) the package and it's transitive dependency switched from CJS to ESM format. This format switch caused a many build failures for consumers using jest (jest can't handle ESM by default, it needs to get transformed).
To avoid breaking consumer builds, we decided to ship d3-scale as bundled dependency.

Bundling is happening as part of the build process. First, TS compile the sources, in a second step, the dependency get's bundled with rollup.

Since we use the built artefacts in our unit tests, they will break in case of any changes in those bundled dependencies.