chore: Update set source map's sourceRoot path
Before this change, the sources pointed to the relative path outside of the code package. This is fine during development but points to a wrong path when consuming the built artifacts.
Setting the `sourceRoot` fixes that and points to the correct location in dev mode and when consuming the
artifacts.