fix: Table has same height when last row is selected
Before this change, when selecting the last row on a table in VR
there has been a 1px jump. Caused by the last row which had the same
default border-bottom-height of 2px as all remaining rows.
So when selecting the last row, the padding-bottom got reduced (as for all remaining rows) when selecting to make the available space for increasing the border-bottom-width from 1px to 2px. This caused the jump.

With this change, when selecting the last row it does not adjust the padding-bottom as it’d be needed for all remaining rows. This keeps the table in the same height in VR.