chore: improve token-group focus handling
When using the "show more" link on token groups, the focus should
be moved to the first token's close button which got visible after clicking.
This makes the token readable to all users and makes screen reader users aware of them.

For doing that, aria-disabled indicates that the remove button is perceivable but disabled.
The onClick callback get's conditionally suppressed depending on the disabled state.