fix: hide popover on line charts when relatedTarget is not a DOM node (#2068)
fix: hide popover on line charts when relatedTarget is not a DOM node
This change addresses the issue where an exception was thrown when leaving
the popover on a chart by either:
- Focusing another element outside the page (all browsers)
- Right-clicking on a popover (Firefox only)

The root cause was that `event.relatedTarget` pointed to the `window` object
in these cases. The `useMouseHover` hook has been updated to account
for this scenario.