feat: Add ariaLabels to token group for "Show fewer" and "Show more" buttons
We have writing guidelines for what text to use for show more/show fewer values, but the visible text and accessible text will have to differ so that the accessible text can be unique across the page.

Added the new properties for Muitlselect and PropertyFilter as well because both use the internal token group component.