feat: Add ariaLabels to token group for "Show fewer" and "Show more" buttons (#2083)
For the TokenGroup component (used in Multiselect and PropertyFilter), we have writing guidelines for the show more/show fewer button labels. When multiple TokenGroups/Multiselect/PropertyFilter components are used on one page with identical show more/show fewer button labels, the button’s aria-label should be different to ensure uniqueness across the page. This enables screen reader users to determine the purpose of the button.