fix: Fix off-center popover flash when clicking CopyToClipboard
Because the clipboard API is async, the popover briefly renders in the
"pending" state before rendering in the "success" or "error" state,
which is perceived as a visual glitch.

The fix here is to wait to open the popover until we're out of the
"pending" state. To achieve that, I've modified the `InternalPopover`
API so that it doesn't render a popover if the given `content` is
`null`, regardless of its internal `visible` state.