chore: Add interactionMode property to InternalDragHandle component
This mode allows the builder to decide when UAP action buttons are shown.

Use-case: On board items the UAP action buttons should be shown on Enter instead of onFocusing. When using onFocus to show them, tabbing through a board with many board items gets visually noisy.

(This is a preparation for adding the InternalDragHandle to board-items)