feat: Use succinct UTC time zone offset: "+00:00" -> "Z"
This is a second attempt at #3763.

This pull request changes how UTC dates are formatted to use the `Z`
suffix instead of `+00:00`. This should make UTC dates more succinct and
readable. cc @jperals

**Before**

```
2018-01-16T00:00:00+00:00
```

**After**

```
2018-01-16T00:00:00Z
```

The key difference between this and #3763 is that it makes a distinction
between formatting ISO timestamps for internal use vs. formatting them
for display. Internal ISO timestamps are used in (for example) the
`onChange` event in `DateRangePicker`. Those internal ISO timestamps
retain the `+00:00` suffix to ensure backward compatibility with
customers' parsing logic.