chore: Ignore unreachable line
Turns out the solution was just to ignore the line because it's unreachable by any test : ). So here's the thing I figured out, if we run the test in ssr window will be undefined but then useLayoutEffect is not invoked in ssr. And on normal client side the useLayoutEffect is invoked but window is defined. Hence technically it's not possible to cover this line with test. So I searched for a way to ignore it and the comment /* istanbul ignore */ came up, and turns out we use it in many other places in our codebase.